requirements.txt
openai==1.63.0
python-dotenv==1.0.1
gTTS==2.5.4
playsound==1.2.2
pygame==2.6.1
spotipy==2.25.1
pytchat==0.5.5
flask==3.0.0
flask-cors==4.0.0